third_party/blink/web_tests/platform/win/external/wpt/selection/drag-selection-extend-to-user-select-none-expected.txt

This is a testharness.js-based test.
[FAIL] Select user-select:text content and then extend selection to the next user-select:text element by crossing the user-select:none element.
  assert_equals: Focus node:  expected Text node " Second: Lorem ipsum dolor sit amet" but got Text node " dolor sit"
Harness: the test ran to completion.
